requirements.txt
absl-py==2.2.2
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
bangla==0.0.2
beautifulsoup4==4.13.3
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
comtypes==1.4.10
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.11
Cython==3.0.12
dateparser==1.1.8
decorator==4.4.2
diffusers==0.32.2
distro==1.9.0
docopt==0.6.2
edge-tts==7.0.0
einops==0.8.1
encodec==0.1.1
exceptiongroup==1.2.2
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
g2pkk==0.1.2
gdown==5.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
googletrans==4.0.0rc1
grpcio==1.71.0
grpcio-status==1.71.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
gTTS==2.5.4
h11==0.9.0
h2==3.2.0
hangul-romanize==0.1.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.0
httplib2==0.22.0
httpx==0.13.3
huggingface-hub==0.30.1
hyperframe==5.2.0
idna==2.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonlines==1.2.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.0
llvmlite==0.44.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==10.6.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
murmurhash==1.0.12
networkx==3.4.2
nltk==3.9.1
noisereduce==3.0.3
num2words==0.5.14
numba==0.61.2
numpy==1.24.3
openai==1.70.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.0.3
pillow==10.4.0
platformdirs==4.3.7
pooch==1.8.2
preshed==3.0.9
proglog==0.1.11
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pygame==2.6.1
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.3
pypinyin==0.54.0
pypiwin32==223
pysbd==0.3.4
PySocks==1.7.1
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyttsx3==2.98
pytz==2025.2
pywin32==310
PyYAML==6.0.2
regex==2024.11.6
replicate==1.0.4
requests==2.32.3
rfc3986==1.5.0
rich==14.0.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.2
srsly==2.5.1
srt==3.5.3
SudachiDict-core==20250129
SudachiPy==0.6.10
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thinc==8.3.4
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
trainer==0.0.36
transformers==4.50.3
translator==0.0.9
TTS==0.22.0
typeguard==4.4.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.7
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.19.0
zipp==3.21.0
